+++ java-io-http/src/test/java/nbbrd/io/http/UriQueryBuilderTest.java
@@ -21,21 +21,20 @@ import org.junit.jupiter.api.Test;
 import wiremock.org.apache.hc.core5.net.URIBuilder;
 
 import java.io.IOException;
-import java.net.MalformedURLException;
+import java.net.URI;
 import java.net.URISyntaxException;
-import java.net.URL;
 import java.util.List;
 import java.util.function.Consumer;
 
 import static java.util.Arrays.asList;
-import static nbbrd.io.http.URLQueryBuilder.of;
+import static nbbrd.io.http.UriQueryBuilder.of;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNullPointerException;
 
 /**
  * @author Philippe Charles
  */
-public class URLQueryBuilderTest {
+public class UriQueryBuilderTest {
 
     @SuppressWarnings("DataFlowIssue")
     @Test
@@ -44,11 +43,11 @@ public class URLQueryBuilderTest {
                 .isThrownBy(() -> of(null));
 
         assertThat(of(urlWithoutSlash).build())
-                .isEqualTo(uriBuilderAsURL(urlWithoutSlash, DO_NOTHING))
+                .isEqualTo(uriBuilderAsURI(urlWithoutSlash, DO_NOTHING))
                 .hasToString("http://localhost");
 
         assertThat(of(urlWithSlash).build())
-                .isEqualTo(uriBuilderAsURL(urlWithSlash, DO_NOTHING))
+                .isEqualTo(uriBuilderAsURI(urlWithSlash, DO_NOTHING))
                 .hasToString("http://localhost/");
     }
 
@@ -59,11 +58,11 @@ public class URLQueryBuilderTest {
                 .isThrownBy(() -> of(urlWithoutSlash).path((String) null));
 
         assertThat(of(urlWithoutSlash).path("hello").path("").path("worl/d").build())
-                .isEqualTo(uriBuilderAsURL(urlWithoutSlash, o -> o.appendPathSegments("hello", "", "worl/d")))
+                .isEqualTo(uriBuilderAsURI(urlWithoutSlash, o -> o.appendPathSegments("hello", "", "worl/d")))
                 .hasToString("http://localhost/hello//worl%2Fd");
 
         assertThat(of(urlWithSlash).path("hello").path("").path("worl/d").build())
-                .isNotEqualTo(uriBuilderAsURL(urlWithSlash, o -> o.appendPathSegments("hello", "", "worl/d")))
+                .isNotEqualTo(uriBuilderAsURI(urlWithSlash, o -> o.appendPathSegments("hello", "", "worl/d")))
                 .hasToString("http://localhost/hello//worl%2Fd");
     }
 
@@ -74,11 +73,11 @@ public class URLQueryBuilderTest {
                 .isThrownBy(() -> of(urlWithoutSlash).path((List<String>) null));
 
         assertThat(of(urlWithoutSlash).path(asList("hello", "", "worl/d")).build())
-                .isEqualTo(uriBuilderAsURL(urlWithoutSlash, o -> o.appendPathSegments("hello", "", "worl/d")))
+                .isEqualTo(uriBuilderAsURI(urlWithoutSlash, o -> o.appendPathSegments("hello", "", "worl/d")))
                 .hasToString("http://localhost/hello//worl%2Fd");
 
         assertThat(of(urlWithSlash).path(asList("hello", "", "worl/d")).build())
-                .isNotEqualTo(uriBuilderAsURL(urlWithSlash, o -> o.appendPathSegments("hello", "", "worl/d")))
+                .isNotEqualTo(uriBuilderAsURI(urlWithSlash, o -> o.appendPathSegments("hello", "", "worl/d")))
                 .hasToString("http://localhost/hello//worl%2Fd");
     }
 
@@ -92,27 +91,27 @@ public class URLQueryBuilderTest {
                 .isThrownBy(() -> of(urlWithoutSlash).param("", null));
 
         assertThat(of(urlWithoutSlash).param("p1", "v1").param("p&=2", "v&=2").build())
-                .isEqualTo(uriBuilderAsURL(urlWithoutSlash, o -> o.addParameter("p1", "v1").addParameter("p&=2", "v&=2")))
+                .isEqualTo(uriBuilderAsURI(urlWithoutSlash, o -> o.addParameter("p1", "v1").addParameter("p&=2", "v&=2")))
                 .hasToString("http://localhost?p1=v1&p%26%3D2=v%26%3D2");
 
         assertThat(of(urlWithSlash).param("p1", "v1").param("p&=2", "v&=2").build())
-                .isEqualTo(uriBuilderAsURL(urlWithSlash, o -> o.addParameter("p1", "v1").addParameter("p&=2", "v&=2")))
+                .isEqualTo(uriBuilderAsURI(urlWithSlash, o -> o.addParameter("p1", "v1").addParameter("p&=2", "v&=2")))
                 .hasToString("http://localhost/?p1=v1&p%26%3D2=v%26%3D2");
 
         assertThat(of(urlWithoutSlash).path("hello").path("worl/d").param("p1", "v1").param("p&=2", "v&=2").build())
-                .isEqualTo(uriBuilderAsURL(urlWithoutSlash, o -> o.appendPathSegments("hello", "worl/d").addParameter("p1", "v1").addParameter("p&=2", "v&=2")))
+                .isEqualTo(uriBuilderAsURI(urlWithoutSlash, o -> o.appendPathSegments("hello", "worl/d").addParameter("p1", "v1").addParameter("p&=2", "v&=2")))
                 .hasToString("http://localhost/hello/worl%2Fd?p1=v1&p%26%3D2=v%26%3D2");
 
         assertThat(of(urlWithSlash).path("hello").path("worl/d").param("p1", "v1").param("p&=2", "v&=2").build())
-                .isNotEqualTo(uriBuilderAsURL(urlWithSlash, o -> o.appendPathSegments("hello", "worl/d").addParameter("p1", "v1").addParameter("p&=2", "v&=2")))
+                .isNotEqualTo(uriBuilderAsURI(urlWithSlash, o -> o.appendPathSegments("hello", "worl/d").addParameter("p1", "v1").addParameter("p&=2", "v&=2")))
                 .hasToString("http://localhost/hello/worl%2Fd?p1=v1&p%26%3D2=v%26%3D2");
 
         assertThat(of(urlWithoutSlash).param("b", "2").param("a", "1").build())
-                .isEqualTo(uriBuilderAsURL(urlWithoutSlash, o -> o.addParameter("b", "2").addParameter("a", "1")))
+                .isEqualTo(uriBuilderAsURI(urlWithoutSlash, o -> o.addParameter("b", "2").addParameter("a", "1")))
                 .hasToString("http://localhost?b=2&a=1");
 
         assertThat(of(urlWithSlash).param("b", "2").param("a", "1").build())
-                .isEqualTo(uriBuilderAsURL(urlWithSlash, o -> o.addParameter("b", "2").addParameter("a", "1")))
+                .isEqualTo(uriBuilderAsURI(urlWithSlash, o -> o.addParameter("b", "2").addParameter("a", "1")))
                 .hasToString("http://localhost/?b=2&a=1");
     }
 
@@ -123,11 +122,11 @@ public class URLQueryBuilderTest {
                 .isThrownBy(() -> of(urlWithoutSlash).param(null));
 
         assertThat(of(urlWithoutSlash).param("b").param("a").build())
-                .isEqualTo(uriBuilderAsURL(urlWithoutSlash, o -> o.addParameter("b", null).addParameter("a", null)))
+                .isEqualTo(uriBuilderAsURI(urlWithoutSlash, o -> o.addParameter("b", null).addParameter("a", null)))
                 .hasToString("http://localhost?b&a");
 
         assertThat(of(urlWithSlash).param("b").param("a").build())
-                .isEqualTo(uriBuilderAsURL(urlWithSlash, o -> o.addParameter("b", null).addParameter("a", null)))
+                .isEqualTo(uriBuilderAsURI(urlWithSlash, o -> o.addParameter("b", null).addParameter("a", null)))
                 .hasToString("http://localhost/?b&a");
     }
 
@@ -149,31 +148,31 @@ public class URLQueryBuilderTest {
     @Test
     public void testEncodingOfSpaceCharacter() throws IOException {
         assertThat(of(urlWithoutSlash).path("a b+").build())
-                .isEqualTo(uriBuilderAsURL(urlWithoutSlash, o -> o.appendPathSegments("a b+")))
+                .isEqualTo(uriBuilderAsURI(urlWithoutSlash, o -> o.appendPathSegments("a b+")))
                 .hasToString("http://localhost/a%20b%2B");
 
         assertThat(of(urlWithSlash).path("a b+").build())
-                .isNotEqualTo(uriBuilderAsURL(urlWithSlash, o -> o.appendPathSegments("a b+")))
+                .isNotEqualTo(uriBuilderAsURI(urlWithSlash, o -> o.appendPathSegments("a b+")))
                 .hasToString("http://localhost/a%20b%2B");
 
         assertThat(of(urlWithoutSlash).param("x y+", "a b+").build())
-                .isEqualTo(uriBuilderAsURL(urlWithoutSlash, o -> o.addParameter("x y+","a b+")))
+                .isEqualTo(uriBuilderAsURI(urlWithoutSlash, o -> o.addParameter("x y+", "a b+")))
                 .hasToString("http://localhost?x%20y%2B=a%20b%2B");
 
         assertThat(of(urlWithSlash).param("x y+", "a b+").build())
-                .isEqualTo(uriBuilderAsURL(urlWithSlash, o -> o.addParameter("x y+","a b+")))
+                .isEqualTo(uriBuilderAsURI(urlWithSlash, o -> o.addParameter("x y+", "a b+")))
                 .hasToString("http://localhost/?x%20y%2B=a%20b%2B");
     }
 
-    private final URL urlWithoutSlash = Parser.onURL().parseValue("http://localhost").orElseThrow(RuntimeException::new);
-    private final URL urlWithSlash = Parser.onURL().parseValue("http://localhost/").orElseThrow(RuntimeException::new);
+    private final URI urlWithoutSlash = Parser.onURI().parseValue("http://localhost").orElseThrow(RuntimeException::new);
+    private final URI urlWithSlash = Parser.onURI().parseValue("http://localhost/").orElseThrow(RuntimeException::new);
 
-    private static URL uriBuilderAsURL(URL base, Consumer<? super URIBuilder> consumer) {
+    private static URI uriBuilderAsURI(URI base, Consumer<? super URIBuilder> consumer) {
         try {
-            URIBuilder builder = new URIBuilder(base.toURI());
+            URIBuilder builder = new URIBuilder(base);
             consumer.accept(builder);
-            return builder.build().toURL();
-        } catch (URISyntaxException | MalformedURLException ex) {
+            return builder.build();
+        } catch (URISyntaxException ex) {
             throw new RuntimeException(ex);
         }
     }